refactor: move schedule_spare_time_input file to schedule_spare_time folder
